requirements.txt
ultralytics>=8.0.0
geopandas>=0.14.0
shapely>=2.0.0
tqdm>=4.66.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
folium>=0.14.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
plotly>=5.17.0

